test: delete bootstrap integration tests, move uncovered tests
Move 6 executeCommand/selectFolder tests to ipc-event-bridge.integration.test.ts
(the module where those handlers are registered) and delete the 1556-line
bootstrap.integration.test.ts whose remaining 24 tests duplicate coverage
from individual module/operation test files.
